#1208 Move the Directive node tests into its own files
