+++ src/psyclone/tests/psyGen_test.py
@@ -35,13 +35,11 @@
 # Modified I. Kavcic, Met Office
 # -----------------------------------------------------------------------------
 
-''' Performs py.test tests on the psygen module '''
+''' Performs py.test tests on the psyGen module '''
 
 
 # internal classes requiring tests
-# PSy,Invokes,Dependencies,Invoke,Node,Schedule,
-# LoopDirective,OMPLoopDirective,Loop,Call,Inf,SetInfCall,Kern,Arguments,
-# InfArguments,Argument,KernelArgument,InfArgument
+# PSy,Invokes,Invoke, Kern,Arguments, Argument, KernelArgument
 
 # user classes requiring tests
 # PSyFactory, TransInfo, Transformation
@@ -50,7 +48,6 @@ import os
 import pytest
 
 from fparser import api as fpapi, logging
-from fparser.common.readfortran import FortranStringReader
 from fparser.two import Fortran2003
 
 from psyclone.configuration import Config
@@ -68,9 +65,7 @@ from psyclone.psyGen import TransInfo, Transformation, PSyFactory, \
     DataAccess, Kern, Arguments, CodedKern, Argument, GlobalSum, \
     InvokeSchedule
 from psyclone.psyir.nodes import Assignment, BinaryOperation, Container, \
-    Literal, Node, Schedule, KernelSchedule, Call, Loop, colored, \
-    OMPParallelDoDirective, ACCEnterDataDirective, ACCKernelsDirective, \
-    OMPParallelDirective, OMPDoDirective, OMPDirective, Directive
+    Literal, Node, KernelSchedule, Call, Loop, colored
 from psyclone.psyir.symbols import DataSymbol, RoutineSymbol, REAL_TYPE, \
     GlobalInterface, ContainerSymbol, Symbol, INTEGER_TYPE, DeferredType, \
     SymbolTable
@@ -78,11 +73,9 @@ from psyclone.tests.lfric_build import LFRicBuild
 from psyclone.tests.test_files import dummy_transformations
 from psyclone.tests.test_files.dummy_transformations import LocalTransformation
 from psyclone.tests.utilities import get_invoke
-from psyclone.transformations import OMPParallelLoopTrans, \
-    Dynamo0p3RedundantComputationTrans, Dynamo0p3KernelConstTrans, \
-    ACCEnterDataTrans, ACCParallelTrans, ACCLoopTrans, ACCKernelsTrans, \
-    Dynamo0p3OMPLoopTrans, OMPParallelTrans, DynamoOMPParallelLoopTrans, \
-    Dynamo0p3ColourTrans
+from psyclone.transformations import Dynamo0p3RedundantComputationTrans, \
+    Dynamo0p3KernelConstTrans, Dynamo0p3OMPLoopTrans, Dynamo0p3ColourTrans, \
+    OMPParallelTrans
 
 
 BASE_PATH = os.path.join(os.path.dirname(os.path.abspath(__file__)),
@@ -805,107 +798,6 @@ def test_kern_is_coloured2():
         assert my_kern.is_coloured()
         loop.loop_type = ""
     assert not my_kern.is_coloured()
-
-
-def test_ompdo_constructor():
-    ''' Check that we can make an OMPDoDirective with and without
-    children '''
-    _, invoke_info = parse(os.path.join(BASE_PATH, "1_single_invoke.f90"),
-                           api="dynamo0.3")
-    psy = PSyFactory("dynamo0.3", distributed_memory=False).create(invoke_info)
-    schedule = psy.invokes.invoke_list[0].schedule
-    ompdo = OMPDoDirective(parent=schedule)
-    # A Directive always has a Schedule
-    assert len(ompdo.children) == 1
-    assert isinstance(ompdo.children[0], Schedule)
-    # Check the dir_body property
-    assert isinstance(ompdo.dir_body, Schedule)
-    # Break the directive
-    del ompdo.children[0]
-    with pytest.raises(InternalError) as err:
-        # pylint: disable=pointless-statement
-        ompdo.dir_body
-    assert ("malformed or incomplete. It should have a single Schedule as a "
-            "child but found: []" in str(err.value))
-    child = schedule.children[0].detach()
-    ompdo = OMPDoDirective(parent=schedule, children=[child])
-    assert len(ompdo.dir_body.children) == 1
-
-
-def test_ompdo_directive_class_node_str(dist_mem):
-    '''Tests the node_str method in the OMPDoDirective class. We create a
-    sub-class object then call this method from it.
-
-    '''
-    _, invoke_info = parse(os.path.join(BASE_PATH, "1_single_invoke.f90"),
-                           api="dynamo0.3")
-
-    cases = [
-        {"current_class": OMPParallelDoDirective,
-         "current_string": "[OMP parallel do]"},
-        {"current_class": OMPDoDirective, "current_string": "[OMP do]"},
-        {"current_class": OMPParallelDirective,
-         "current_string": "[OMP parallel]"},
-        {"current_class": OMPDirective, "current_string": "[OMP]"},
-        {"current_class": Directive, "current_string": ""}]
-    otrans = OMPParallelLoopTrans()
-
-    psy = PSyFactory("dynamo0.3", distributed_memory=dist_mem).\
-        create(invoke_info)
-    schedule = psy.invokes.invoke_list[0].schedule
-
-    if dist_mem:
-        idx = 4
-    else:
-        idx = 0
-
-    _, _ = otrans.apply(schedule.children[idx])
-    omp_parallel_loop = schedule.children[idx]
-
-    for case in cases:
-        # Call the OMPDirective node_str method
-        out = case["current_class"].node_str(omp_parallel_loop)
-
-        directive = colored("Directive", Directive._colour)
-        expected_output = directive + case["current_string"]
-
-        assert expected_output in out
-
-
-def test_acc_dir_node_str():
-    ''' Test the node_str() method of OpenACC directives '''
-
-    acclt = ACCLoopTrans()
-    accdt = ACCEnterDataTrans()
-    accpt = ACCParallelTrans()
-    _, invoke = get_invoke("single_invoke.f90", "gocean1.0", idx=0,
-                           dist_mem=False)
-    colour = Directive._colour
-    schedule = invoke.schedule
-
-    # Enter-data
-    accdt.apply(schedule)
-    out = schedule[0].node_str()
-    assert out.startswith(
-        colored("Directive", colour)+"[ACC enter data]")
-
-    # Parallel region around outermost loop
-    accpt.apply(schedule[1])
-    out = schedule[1].node_str()
-    assert out.startswith(
-        colored("Directive", colour)+"[ACC Parallel]")
-
-    # Loop directive on outermost loop
-    acclt.apply(schedule[1].dir_body[0])
-    out = schedule[1].dir_body[0].node_str()
-    assert out.startswith(
-        colored("Directive", colour)+"[ACC Loop, independent]")
-
-    # Loop directive with collapse
-    acclt.apply(schedule[1].dir_body[0].dir_body[0], {"collapse": 2})
-    out = schedule[1].dir_body[0].dir_body[0].node_str()
-    assert out.startswith(
-        colored("Directive", colour) + "[ACC Loop, collapse=2, independent]")
 
 
 def test_haloexchange_unknown_halo_depth():
@@ -1641,457 +1533,6 @@ def test_call_backward_dependence():
     assert last_call_node.backward_dependence() == prev_dep_call_node
     # b) previous
     assert prev_dep_call_node.backward_dependence() == call3
-
-
-def test_omp_forward_dependence():
-    '''Test that the forward_dependence method works for Directives,
-    returning the closest dependent Node after the current Node in the
-    schedule or None if none are found. '''
-    _, invoke_info = parse(
-        os.path.join(BASE_PATH, "15.14.1_multi_aX_plus_Y_builtin.f90"),
-        api="dynamo0.3")
-    psy = PSyFactory("dynamo0.3", distributed_memory=True).create(invoke_info)
-    invoke = psy.invokes.invoke_list[0]
-    schedule = invoke.schedule
-    otrans = DynamoOMPParallelLoopTrans()
-    for child in schedule.children:
-        otrans.apply(child)
-    read4 = schedule.children[4]
-    # 1: returns none if none found
-    # a) check many reads
-    assert not read4.forward_dependence()
-    # b) check no dependencies for the loop
-    assert not read4.children[0].forward_dependence()
-    # 2: returns first dependent kernel arg when there are many
-    # dependencies
-    # a) check first read returned
-    writer = schedule.children[3]
-    next_read = schedule.children[4]
-    assert writer.forward_dependence() == next_read
-    # b) check writer returned
-    first_omp = schedule.children[0]
-    assert first_omp.forward_dependence() == writer
-    # 3: directive and globalsum dependencies
-    _, invoke_info = parse(
-        os.path.join(BASE_PATH, "15.14.3_sum_setval_field_builtin.f90"),
-        api="dynamo0.3")
-    psy = PSyFactory("dynamo0.3", distributed_memory=True).create(invoke_info)
-    invoke = psy.invokes.invoke_list[0]
-    schedule = invoke.schedule
-    otrans.apply(schedule.children[0])
-    otrans.apply(schedule.children[1])
-    otrans.apply(schedule.children[3])
-    prev_omp = schedule.children[0]
-    sum_omp = schedule.children[1]
-    global_sum_loop = schedule.children[2]
-    next_omp = schedule.children[3]
-    # a) prev omp depends on sum omp
-    assert prev_omp.forward_dependence() == sum_omp
-    # b) sum omp depends on global sum loop
-    assert sum_omp.forward_dependence() == global_sum_loop
-    # c) global sum loop depends on next omp
-    assert global_sum_loop.forward_dependence() == next_omp
-
-
-def test_directive_backward_dependence():
-    '''Test that the backward_dependence method works for Directives,
-    returning the closest dependent Node before the current Node in
-    the schedule or None if none are found.'''
-    _, invoke_info = parse(
-        os.path.join(BASE_PATH, "15.14.1_multi_aX_plus_Y_builtin.f90"),
-        api="dynamo0.3")
-    psy = PSyFactory("dynamo0.3", distributed_memory=True).create(invoke_info)
-    invoke = psy.invokes.invoke_list[0]
-    schedule = invoke.schedule
-    otrans = DynamoOMPParallelLoopTrans()
-    for child in schedule.children:
-        otrans.apply(child)
-    # 1: omp directive no backwards dependence
-    omp3 = schedule.children[2]
-    assert not omp3.backward_dependence()
-    # 2: omp to omp backward dependence
-    # a) many steps
-    last_omp_node = schedule.children[6]
-    prev_dep_omp_node = schedule.children[3]
-    assert last_omp_node.backward_dependence() == prev_dep_omp_node
-    # b) previous
-    assert prev_dep_omp_node.backward_dependence() == omp3
-    # 3: globalsum dependencies
-    _, invoke_info = parse(
-        os.path.join(BASE_PATH, "15.14.3_sum_setval_field_builtin.f90"),
-        api="dynamo0.3")
-    psy = PSyFactory("dynamo0.3", distributed_memory=True).create(invoke_info)
-    invoke = psy.invokes.invoke_list[0]
-    schedule = invoke.schedule
-    otrans.apply(schedule.children[0])
-    otrans.apply(schedule.children[1])
-    otrans.apply(schedule.children[3])
-    omp1 = schedule.children[0]
-    omp2 = schedule.children[1]
-    global_sum = schedule.children[2]
-    omp3 = schedule.children[3]
-    # a) omp3 depends on global sum
-    assert omp3.backward_dependence() == global_sum
-    # b) global sum depends on omp2
-    assert global_sum.backward_dependence() == omp2
-    # c) omp2 (sum) depends on omp1
-    assert omp2.backward_dependence() == omp1
-
-
-def test_directive_get_private(monkeypatch):
-    ''' Tests for the _get_private_list() method of OMPParallelDirective.
-    Note: this test does not apply colouring so the loops must be over
-    discontinuous function spaces.
-
-    '''
-    _, invoke_info = parse(
-        os.path.join(BASE_PATH, "1_single_invoke_w3.f90"), api="dynamo0.3")
-    psy = PSyFactory("dynamo0.3",
-                     distributed_memory=False).create(invoke_info)
-    invoke = psy.invokes.invoke_list[0]
-    schedule = invoke.schedule
-    # We use Transformations to introduce the necessary directives
-    otrans = Dynamo0p3OMPLoopTrans()
-    rtrans = OMPParallelTrans()
-    # Apply an OpenMP do directive to the loop
-    otrans.apply(schedule.children[0], {"reprod": True})
-    # Apply an OpenMP Parallel directive around the OpenMP do directive
-    rtrans.apply(schedule.children[0])
-    directive = schedule.children[0]
-    assert isinstance(directive, OMPParallelDirective)
-    # Now check that _get_private_list returns what we expect
-    pvars = directive._get_private_list()
-    assert pvars == ['cell']
-    # Now use monkeypatch to break the Call within the loop
-    call = directive.dir_body[0].dir_body[0].loop_body[0]
-    monkeypatch.setattr(call, "local_vars", lambda: [""])
-    with pytest.raises(InternalError) as err:
-        _ = directive._get_private_list()
-    assert ("call 'testkern_w3_code' has a local variable but its name is "
-            "not set" in str(err.value))
-
-
-def test_directive_children_validation():
-    '''Test that children added to Directive are validated. Directive accepts
-    1 Schedule as child.
-
-    '''
-    directive = Directive()
-    datanode = Literal("1", INTEGER_TYPE)
-    schedule = Schedule()
-
-    # First child
-    with pytest.raises(GenerationError) as excinfo:
-        directive.children[0] = datanode
-    assert ("Item 'Literal' can't be child 0 of 'Directive'. The valid format"
-            " is: 'Schedule'." in str(excinfo.value))
-
-    # Additional children
-    with pytest.raises(GenerationError) as excinfo:
-        directive.addchild(schedule)
-    assert ("Item 'Schedule' can't be child 1 of 'Directive'. The valid format"
-            " is: 'Schedule'." in str(excinfo.value))
-
-
-def test_openmp_pdo_dag_name():
-    '''Test that we generate the correct dag name for the OpenMP parallel
-    do node'''
-    _, info = parse(os.path.join(BASE_PATH,
-                                 "15.7.2_setval_X_builtin.f90"),
-                    api="dynamo0.3")
-    psy = PSyFactory("dynamo0.3", distributed_memory=False).create(info)
-    invoke = psy.invokes.invoke_list[0]
-    schedule = invoke.schedule
-    otrans = DynamoOMPParallelLoopTrans()
-    # Apply OpenMP parallelisation to the loop
-    otrans.apply(schedule.children[0])
-    assert schedule.children[0].dag_name == "OMP_parallel_do_1"
-
-
-def test_omp_dag_names():
-    ''' Test that we generate the correct dag names for omp parallel, omp
-    do, omp directive and directive nodes.
-    Note: this test does not apply colouring so the loops must be over
-    discontinuous function spaces.
-
-    '''
-    _, info = parse(os.path.join(os.path.dirname(os.path.abspath(__file__)),
-                                 "test_files", "dynamo0p3",
-                                 "1_single_invoke_w3.f90"),
-                    api="dynamo0.3")
-    psy = PSyFactory("dynamo0.3", distributed_memory=False).create(info)
-    invoke = psy.invokes.get('invoke_0_testkern_w3_type')
-    schedule = invoke.schedule
-    olooptrans = Dynamo0p3OMPLoopTrans()
-    ptrans = OMPParallelTrans()
-    # Put an OMP PARALLEL around this loop
-    child = schedule.children[0]
-    ptrans.apply(child)
-    # Put an OMP DO around this loop
-    olooptrans.apply(schedule[0].dir_body[0])
-    omp_par_node = schedule.children[0]
-    assert omp_par_node.dag_name == "OMP_parallel_1"
-    assert omp_par_node.dir_body[0].dag_name == "OMP_do_3"
-    omp_directive = super(OMPParallelDirective, omp_par_node)
-    assert omp_directive.dag_name == "OMP_directive_1"
-    directive = super(OMPDirective, omp_par_node)
-    assert directive.dag_name == "directive_1"
-
-
-def test_acc_dag_names():
-    ''' Check that we generate the correct dag names for ACC parallel,
-    ACC enter-data and ACC loop directive Nodes '''
-    _, invoke = get_invoke("single_invoke.f90", "gocean1.0", idx=0,
-                           dist_mem=False)
-    schedule = invoke.schedule
-
-    acclt = ACCLoopTrans()
-    accdt = ACCEnterDataTrans()
-    accpt = ACCParallelTrans()
-    # Enter-data
-    accdt.apply(schedule)
-    assert schedule[0].dag_name == "ACC_data_1"
-    # Parallel region
-    accpt.apply(schedule[1])
-    assert schedule[1].dag_name == "ACC_parallel_3"
-    # Loop directive
-    acclt.apply(schedule[1].dir_body[0])
-    assert schedule[1].dir_body[0].dag_name == "ACC_loop_5"
-    # Base class
-    name = super(ACCEnterDataDirective, schedule[0]).dag_name
-    assert name == "ACC_directive_1"
-
-# Class ACCKernelsDirective start
-
-
-# (1/1) Method __init__
-def test_acckernelsdirective_init():
-    '''Test an ACCKernelsDirective can be created and that the optional
-    arguments are set and can be set as expected.
-
-    '''
-    directive = ACCKernelsDirective()
-    assert directive._default_present
-    assert directive.parent is None
-    assert len(directive.children) == 1
-    assert isinstance(directive.children[0], Schedule)
-    directive = ACCKernelsDirective(default_present=False)
-    assert not directive._default_present
-
-
-# (1/1) Method dag_name
-def test_acckernelsdirective_dagname():
-    '''Check that the dag_name method in the ACCKernelsDirective class
-    behaves as expected.
-
-    '''
-    _, info = parse(os.path.join(BASE_PATH, "1_single_invoke.f90"))
-    psy = PSyFactory(distributed_memory=False).create(info)
-    sched = psy.invokes.get('invoke_0_testkern_type').schedule
-
-    trans = ACCKernelsTrans()
-    _, _ = trans.apply(sched)
-    assert sched.children[0].dag_name == "ACC_kernels_1"
-
-
-# (1/1) Method node_str
-def test_acckernelsdirective_node_str():
-    '''Check that the node_str method in the ACCKernelsDirective class behaves
-    as expected.
-
-    '''
-    _, info = parse(os.path.join(BASE_PATH, "1_single_invoke.f90"))
-    psy = PSyFactory(distributed_memory=False).create(info)
-    sched = psy.invokes.get('invoke_0_testkern_type').schedule
-
-    trans = ACCKernelsTrans()
-    _, _ = trans.apply(sched)
-
-    out = sched[0].node_str()
-    assert out.startswith(
-        colored("Directive", Directive._colour)+"[ACC Kernels]")
-    assert colored("Loop", Loop._colour) in sched[0].dir_body[0].node_str()
-    assert "CodedKern" in sched[0].dir_body[0].loop_body[0].node_str()
-
-
-# (1/1) Method gen_code
-@pytest.mark.parametrize("default_present", [False, True])
-def test_acckernelsdirective_gencode(default_present):
-    '''Check that the gen_code method in the ACCKernelsDirective class
-    generates the expected code. Use the dynamo0.3 API.
-
-    '''
-    _, info = parse(os.path.join(BASE_PATH, "1_single_invoke.f90"))
-    psy = PSyFactory(distributed_memory=False).create(info)
-    sched = psy.invokes.get('invoke_0_testkern_type').schedule
-
-    trans = ACCKernelsTrans()
-    _, _ = trans.apply(sched, {"default_present": default_present})
-
-    code = str(psy.gen)
-    string = ""
-    if default_present:
-        string = " default(present)"
-    assert (
-        "      !$acc kernels{0}\n"
-        "      DO cell=1,f1_proxy%vspace%get_ncell()\n".format(string) in code)
-    assert (
-        "      END DO\n"
-        "      !$acc end kernels\n" in code)
-
-
-# (1/1) Method update
-@pytest.mark.parametrize("default_present", [False, True])
-def test_acckernelsdirective_update(parser, default_present):
-    '''Check that the update method in the ACCKernelsDirective class
-    generates the expected code. Use the nemo API.
-
-    '''
-    reader = FortranStringReader("program implicit_loop\n"
-                                 "real(kind=wp) :: sto_tmp(5,5)\n"
-                                 "sto_tmp(:,:) = 0.0_wp\n"
-                                 "end program implicit_loop\n")
-    code = parser(reader)
-    psy = PSyFactory("nemo", distributed_memory=False).create(code)
-    schedule = psy.invokes.invoke_list[0].schedule
-    kernels_trans = ACCKernelsTrans()
-    kernels_trans.apply(schedule.children[0:1],
-                        {"default_present": default_present})
-    gen_code = str(psy.gen)
-    string = ""
-    if default_present:
-        string = " DEFAULT(PRESENT)"
-    assert ("  !$ACC KERNELS{0}\n"
-            "  sto_tmp(:, :) = 0.0_wp\n"
-            "  !$ACC END KERNELS\n".format(string) in gen_code)
-
-# Class ACCKernelsDirective end
-
-# Class ACCEnterDataDirective start
-
-
-# (1/1) Method __init__
-def test_acc_datadevice_virtual():
-    ''' Check that we can't instantiate an instance of
-    ACCEnterDataDirective. '''
-    # pylint:disable=abstract-class-instantiated
-    with pytest.raises(TypeError) as err:
-        ACCEnterDataDirective()
-    # pylint:enable=abstract-class-instantiated
-    assert ("instantiate abstract class ACCEnterDataDirective with abstract "
-            "methods data_on_device" in str(err.value))
-
-# (1/1) Method node_str
-# Covered in test test_acc_dir_node_str
-
-# (1/1) Method dag_name
-# Covered in test_acc_dag_names
-
-
-# (1/4) Method gen_code
-def test_accenterdatadirective_gencode_1():
-    '''Test that an OpenACC Enter Data directive, when added to a schedule
-    with a single loop, raises the expected exception as there is no
-    following OpenACC Parallel or OpenACC Kernels directive as at
-    least one is required. This test uses the dynamo0.3 API.
-
-    '''
-    acc_enter_trans = ACCEnterDataTrans()
-    _, info = parse(os.path.join(BASE_PATH, "1_single_invoke.f90"))
-    psy = PSyFactory(distributed_memory=False).create(info)
-    sched = psy.invokes.get('invoke_0_testkern_type').schedule
-    acc_enter_trans.apply(sched)
-    with pytest.raises(GenerationError) as excinfo:
-        str(psy.gen)
-    assert ("ACCEnterData directive did not find any data to copyin. Perhaps "
-            "there are no ACCParallel or ACCKernels directives within the "
-            "region." in str(excinfo.value))
-
-    # Test that the same error is produced by the begin_string() which is used
-    # by the PSyIR backend
-    sched[0].lower_to_language_level()
-    with pytest.raises(GenerationError) as excinfo:
-        sched[0].begin_string()
-    assert ("ACCEnterData directive did not find any data to copyin. Perhaps "
-            "there are no ACCParallel or ACCKernels directives within the "
-            "region." in str(excinfo.value))
-
-
-# (2/4) Method gen_code
-def test_accenterdatadirective_gencode_2():
-    '''Test that an OpenACC Enter Data directive, when added to a schedule
-    with multiple loops, raises the expected exception, as there is no
-    following OpenACC Parallel or OpenACCKernels directive and at
-    least one is required. This test uses the dynamo0.3 API.
-
-    '''
-    acc_enter_trans = ACCEnterDataTrans()
-    _, info = parse(os.path.join(BASE_PATH, "1.2_multi_invoke.f90"))
-    psy = PSyFactory(distributed_memory=False).create(info)
-    sched = psy.invokes.get('invoke_0').schedule
-    acc_enter_trans.apply(sched)
-    with pytest.raises(GenerationError) as excinfo:
-        str(psy.gen)
-    assert ("ACCEnterData directive did not find any data to copyin. Perhaps "
-            "there are no ACCParallel or ACCKernels directives within the "
-            "region." in str(excinfo.value))
-
-
-# (3/4) Method gen_code
-@pytest.mark.parametrize("trans", [ACCParallelTrans, ACCKernelsTrans])
-def test_accenterdatadirective_gencode_3(trans):
-    '''Test that an OpenACC Enter Data directive, when added to a schedule
-    with a single loop, produces the expected code (there should be
-    "copy in" data as there is a following OpenACC parallel or kernels
-    directive). This test uses the dynamo0.3 API.
-
-    '''
-    acc_trans = trans()
-    acc_enter_trans = ACCEnterDataTrans()
-    _, info = parse(os.path.join(BASE_PATH, "1_single_invoke.f90"))
-    psy = PSyFactory(distributed_memory=False).create(info)
-    sched = psy.invokes.get('invoke_0_testkern_type').schedule
-    _ = acc_trans.apply(sched.children)
-    _ = acc_enter_trans.apply(sched)
-    code = str(psy.gen)
-    assert (
-        "      !$acc enter data copyin(nlayers,a,f1_proxy,f1_proxy%data,"
-        "f2_proxy,f2_proxy%data,m1_proxy,m1_proxy%data,m2_proxy,"
-        "m2_proxy%data,ndf_w1,undf_w1,map_w1,ndf_w2,undf_w2,map_w2,"
-        "ndf_w3,undf_w3,map_w3)\n" in code)
-
-
-# (4/4) Method gen_code
-@pytest.mark.parametrize("trans1,trans2",
-                         [(ACCParallelTrans, ACCParallelTrans),
-                          (ACCParallelTrans, ACCKernelsTrans),
-                          (ACCKernelsTrans, ACCParallelTrans),
-                          (ACCKernelsTrans, ACCKernelsTrans)])
-def test_accenterdatadirective_gencode_4(trans1, trans2):
-    '''Test that an OpenACC Enter Data directive, when added to a schedule
-    with multiple loops and multiple OpenACC parallel and/or Kernel
-    directives, produces the expected code (when the same argument is
-    used in multiple loops there should only be one entry). This test
-    uses the dynamo0.3 API.
-
-    '''
-    acc_trans1 = trans1()
-    acc_trans2 = trans2()
-    acc_enter_trans = ACCEnterDataTrans()
-    _, info = parse(os.path.join(BASE_PATH, "1.2_multi_invoke.f90"))
-    psy = PSyFactory(distributed_memory=False).create(info)
-    sched = psy.invokes.get('invoke_0').schedule
-    _ = acc_trans1.apply([sched.children[1]])
-    _ = acc_trans2.apply([sched.children[0]])
-    _ = acc_enter_trans.apply(sched)
-    code = str(psy.gen)
-    assert (
-        "      !$acc enter data copyin(nlayers,a,f1_proxy,f1_proxy%data,"
-        "f2_proxy,f2_proxy%data,m1_proxy,m1_proxy%data,m2_proxy,m2_proxy%data,"
-        "ndf_w1,undf_w1,map_w1,ndf_w2,undf_w2,map_w2,ndf_w3,undf_w3,map_w3,"
-        "f3_proxy,f3_proxy%data)\n" in code)
-
-# Class ACCEnterDataDirective end
 
 
 def test_haloexchange_halo_depth_get_set():
