+++ src/psyclone/tests/psyir/nodes/directive_test.py
@@ -1,0 +1,118 @@
+# -----------------------------------------------------------------------------
+# BSD 3-Clause License
+#
+# Copyright (c) 2021, Science and Technology Facilities Council.
+# All rights reserved.
+#
+# Redistribution and use in source and binary forms, with or without
+# modification, are permitted provided that the following conditions are met:
+#
+# * Redistributions of source code must retain the above copyright notice, this
+#   list of conditions and the following disclaimer.
+#
+# * Redistributions in binary form must reproduce the above copyright notice,
+#   this list of conditions and the following disclaimer in the documentation
+#   and/or other materials provided with the distribution.
+#
+# * Neither the name of the copyright holder nor the names of its
+#   contributors may be used to endorse or promote products derived from
+#   this software without specific prior written permission.
+#
+# THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
+# "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
+# LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS
+# FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE
+# COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT,
+# INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
+# BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
+# LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER
+# CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT
+# LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN
+# ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
+# POSSIBILITY OF SUCH DAMAGE.
+# -----------------------------------------------------------------------------
+# Authors R. W. Ford, A. R. Porter and S. Siso, STFC Daresbury Lab
+# Modified I. Kavcic, Met Office
+# -----------------------------------------------------------------------------
+
+''' Performs py.test tests on the PSyIR Directive node. '''
+
+from __future__ import absolute_import
+import os
+import pytest
+from psyclone.parse.algorithm import parse
+from psyclone.psyGen import PSyFactory
+from psyclone.psyir.nodes import Directive, Literal, Schedule
+from psyclone.errors import GenerationError
+from psyclone.psyir.symbols import INTEGER_TYPE
+from psyclone.transformations import DynamoOMPParallelLoopTrans
+
+BASE_PATH = os.path.join(os.path.dirname(os.path.dirname(os.path.dirname(
+    os.path.abspath(__file__)))), "test_files", "dynamo0p3")
+
+
+def test_directive_backward_dependence():
+    '''Test that the backward_dependence method works for Directives,
+    returning the closest dependent Node before the current Node in
+    the schedule or None if none are found.'''
+    _, invoke_info = parse(
+        os.path.join(BASE_PATH, "15.14.1_multi_aX_plus_Y_builtin.f90"),
+        api="dynamo0.3")
+    psy = PSyFactory("dynamo0.3", distributed_memory=True).create(invoke_info)
+    invoke = psy.invokes.invoke_list[0]
+    schedule = invoke.schedule
+    otrans = DynamoOMPParallelLoopTrans()
+    for child in schedule.children:
+        otrans.apply(child)
+    # 1: omp directive no backwards dependence
+    omp3 = schedule.children[2]
+    assert not omp3.backward_dependence()
+    # 2: omp to omp backward dependence
+    # a) many steps
+    last_omp_node = schedule.children[6]
+    prev_dep_omp_node = schedule.children[3]
+    assert last_omp_node.backward_dependence() == prev_dep_omp_node
+    # b) previous
+    assert prev_dep_omp_node.backward_dependence() == omp3
+    # 3: globalsum dependencies
+    _, invoke_info = parse(
+        os.path.join(BASE_PATH, "15.14.3_sum_setval_field_builtin.f90"),
+        api="dynamo0.3")
+    psy = PSyFactory("dynamo0.3", distributed_memory=True).create(invoke_info)
+    invoke = psy.invokes.invoke_list[0]
+    schedule = invoke.schedule
+    otrans.apply(schedule.children[0])
+    otrans.apply(schedule.children[1])
+    otrans.apply(schedule.children[3])
+    omp1 = schedule.children[0]
+    omp2 = schedule.children[1]
+    global_sum = schedule.children[2]
+    omp3 = schedule.children[3]
+    # a) omp3 depends on global sum
+    assert omp3.backward_dependence() == global_sum
+    # b) global sum depends on omp2
+    assert global_sum.backward_dependence() == omp2
+    # c) omp2 (sum) depends on omp1
+    assert omp2.backward_dependence() == omp1
+
+
+def test_directive_children_validation():
+    '''Test that children added to Directive are validated. Directive accepts
+    1 Schedule as child.
+
+    '''
+    directive = Directive()
+    datanode = Literal("1", INTEGER_TYPE)
+    schedule = Schedule()
+
+    # First child
+    with pytest.raises(GenerationError) as excinfo:
+        directive.children[0] = datanode
+    assert ("Item 'Literal' can't be child 0 of 'Directive'. The valid format"
+            " is: 'Schedule'." in str(excinfo.value))
+
+    # Additional children
+    with pytest.raises(GenerationError) as excinfo:
+        directive.addchild(schedule)
+    assert ("Item 'Schedule' can't be child 1 of 'Directive'. The valid format"
+            " is: 'Schedule'." in str(excinfo.value))
